Creates HasNoSaves component
